% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{compatibility}
\alias{compatibility}
\title{Compatibility of conditions and models}
\usage{
compatibility(conditions, model)
}
\arguments{
\item{conditions}{\code{data.frame} with \code{time} and variables expected to correspond to those of the model.}

\item{model}{\code{character} corresponding to the name of the rate model.}
}
\value{
Return \code{TRUE} if the conditions and models are compatibles and can be used to scale time, return an error otherwise.
}
\description{
Test that conditions is compatible with a rate model to scale time.
}
\details{
Variables in the model are all function inputs excluding \code{param} and \code{control} and those of conditions are the column names except \code{time}. The function test that:
(1) variable names are the same in the model and conditions;
(2) code{time} is part of column names of conditions but is not a variable of the model;
(3) column names of conditions are not code{param} and \code{control}, as those are reserved terms of the model function;
(4) model has at least one variable.
}
\examples{
conditions <- data.frame(time = seq(1,30,length.out = 10), temp = rnorm(10, 10, 5))
compatibility(conditions, model = "modelLinear")

}
