% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelLinear.R
\name{modelLinear}
\alias{modelLinear}
\title{Linear rate model in function of temperature}
\usage{
modelLinear(temp, param = list(a, T0))
}
\arguments{
\item{temp}{\code{numeric} vector of temperature (in celcius)}

\item{param}{\code{list} of parameters named \code{T0}, the base temperature under which no development occurs, and \code{a} a normalizing factor such as development is completed at 1 (i.e. \code{a = 1/GDD}).}
}
\value{
return a vector of rate associated to each element of \code{temp}.
}
\description{
Correspond to normalized growing degree day model, without daily approximation of temperature.
}
\examples{
modelLinear(temp = seq(0,30,2), param = list(a = 1, T0 = 10))

}
