% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rleInterval.R
\name{rleInterval}
\alias{rleInterval}
\title{rle interval}
\usage{
rleInterval(x, value, n)
}
\arguments{
\item{x}{an atomic vector .}

\item{value}{a single value to find in \code{x}.}

\item{n}{the minimum of repetition to be considered.}
}
\value{
Return a list of two vectors (\code{lower} and \code{upper}) giving the position ofthe beginning and end of the intervals respecting the criteria.
}
\description{
Define intervals limit (\code{lower}, \code{upper}) on which a vector (\code{x}) repeat a specified value at least \code{n} time (without any other values ocuring). The function relies on the \code{rle} function from \code{base}.
}
\examples{
x <- sample(1:3, size = 100, replace = TRUE)
rleInterval(x, value = 1, n = 1)
rleInterval(x, value = 1, n = 2)
rleInterval(x, value = 1, n = 100)

}
