% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timescale-package.R
\docType{package}
\name{timescale-package}
\alias{timescale}
\alias{timescale-package}
\title{timescale: A package for development time calculation and other time scalings}
\description{
The package ...
\cr\cr
More precisely...
\cr\cr
}
\author{
\strong{Maintainer}: Alexandre Leblanc \email{aleblanc.bio@gmail.com}

}
