% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validityConditions.R
\name{validityConditions}
\alias{validityConditions}
\title{Conditions validity}
\usage{
validityConditions(conditions, model)
}
\arguments{
\item{conditions}{\code{data.frame} with \code{time} and variables expected to correspond to those of the model.}

\item{model}{\code{character} corresponding to the name of the rate model.}
}
\value{
Return \code{TRUE} if the conditions are compatibles with the model, return an error otherwise.
}
\description{
Check the validity of conditions with respect to its structure and compatibility with a rate model for scaling time.
}
\details{
Variables in the model are all function inputs excluding \code{param} and \code{control} and those of conditions are the column names except \code{time}. The function test that:
(1) column names include \code{time};
(2) column names exclude \code{param} and \code{control}, as those are reserved terms of the model function;
(3) variable names are the same in the model and conditions.
}
\examples{
conditions <- data.frame(time = seq(1,30,length.out = 10), temp = rnorm(10, 10, 5))
validityConditions(conditions, model = "modelLinear")
}
