% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validityElement.R
\name{validityElement}
\alias{validityElement}
\title{Validate if is element}
\usage{
validityElement(x, y, name)
}
\arguments{
\item{x}{vectors (of the same mode) containing a sequence
    of items (conceptually) with no duplicated values.}

\item{y}{vectors (of the same mode) containing a sequence
    of items (conceptually) with no duplicated values.}

\item{name}{variable name to return in the error message}
}
\value{
Return \code{TRUE} if the condition is respected and an error otherwise.
}
\description{
Check that all values of a vector \code{x} is element of values of \code{y} and return an error otherwise.
}
\examples{
x <- c(1,2,3)
y <- c(2,3)
z <- c(1,2,3,4)
validityElement(x, y, name = "w")
validityElement(x, z, name = "w")

}
