% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validityLength.R
\name{validityLength}
\alias{validityLength}
\title{Validate if proper length}
\usage{
validityLength(x, value, name)
}
\arguments{
\item{x}{an \R object.  For replacement, a vector or factor.}

\item{value}{a non-negative integer or double (which will be rounded down).}

\item{name}{variable name to return in the error message}
}
\value{
Return \code{TRUE} if the condition is respected and an error otherwise.
}
\description{
Check that a vector \code{x} is of specified length (\code{value}) and return an error otherwise.
}
\examples{
x <- c(1,2,3)
validityLength(x, 3, name = "w")
validityLength(x, 4, name = "w")
}
