% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validityModel.R
\name{validityModel}
\alias{validityModel}
\title{Model validity}
\usage{
validityModel(model)
}
\arguments{
\item{model}{\code{character} corresponding to the name of the rate model.}
}
\value{
Return \code{TRUE} if the model is valid, return an error otherwise.
}
\description{
Check the validity of a rate model for scaling time.
}
\details{
Variables in the model are all function inputs excluding \code{param} and \code{control}. The function test that:
(1) code{time} is not a variable of the model;
(2) the model has at least one variable.
}
