% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validityScaledTime.R
\name{validityScaledTime}
\alias{validityScaledTime}
\title{Scaled time validity}
\usage{
validityScaledTime(
  x,
  model,
  conditions,
  param = list(),
  control = list(),
  interpolation = "constant"
)
}
\arguments{
\item{x}{\code{numeric} vector representing scaled time.}

\item{model}{\code{character} corresponding to the name of the rate model.}

\item{conditions}{\code{data.frame} with columns named \code{time} and variables expected to correspond to those of the model.}

\item{param}{\code{list} parameters of the model.}

\item{control}{\code{list} of arguments that control the behaviour of the model.}

\item{interpolation}{\code{character} corresponding to the name of the interpolating method for conditions. Available methods only include \code{constant} at the moment and is the default value.}
}
\value{
Return \code{TRUE} if the condition is respected and an error otherwise.
}
\description{
Check that a scaled time vector (\code{x}) is included in the scaled time range of \code{conditions}.
}
\examples{
conditions <- data.frame(time = seq(0,30,length.out = 10), temp = rnorm(10, 10, 5))
condModel <- interpolateCond(conditions, method = "linear")
model <- "modelLinear"
param = list(a = 1, T0 = 10)
x = seq(1,10,length.out = 10)
scaledPeriod = rep(10,10)
validityScaledTime(x, model = model, conditions = conditions, param = param)
}
