% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validityTime.R
\name{validityTime}
\alias{validityTime}
\title{Time validity}
\usage{
validityTime(x, conditions)
}
\arguments{
\item{x}{\code{numeric} vector representing time.}

\item{conditions}{\code{data.frame} with \code{time} and variables expected to correspond to those of the model.}
}
\value{
Return \code{TRUE} if the condition is respected and an error otherwise.
}
\description{
Check that a time vector (\code{x}), is included in the range of \code{conditions} time.
}
\details{
The check serve to ensure conditions can be represented through interpolation on \code{x}.
}
\examples{
conditions <- data.frame(time = seq(1,30,length.out = 10), temp = rnorm(10, 10, 5))
validityTime(x = 1:10,conditions)
}
