% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableConditions.R
\name{variableConditions}
\alias{variableConditions}
\title{Model variables}
\usage{
variableConditions(conditions)
}
\arguments{
\item{conditions}{\code{data.frame} with \code{time} and variables as other columns.}
}
\value{
Return \code{character} vector made of the variable names.
}
\description{
Access variable from conditions data.frame.
}
\examples{
conditions <- data.frame(time = seq(1,30,length.out = 10), temp = rnorm(10, 10, 5))
variableConditions(conditions)
}
