% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variableModel.R
\name{variableModel}
\alias{variableModel}
\title{Model variables}
\usage{
variableModel(model)
}
\arguments{
\item{model}{\code{character} corresponding to the name of the rate model.}
}
\value{
Return \code{TRUE} if the conditions and models are compatibles and can be used to scale time, return an error otherwise.
}
\description{
Access variable from a rate model function.
}
\details{
Variables in the model are all function inputs excluding \code{param} and \code{control}.
}
