% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectIntegral.R
\name{vectIntegral}
\alias{vectIntegral}
\title{Integral (vectorized)}
\usage{
vectIntegral(f, xmin, xmax, ...)
}
\arguments{
\item{f}{a function to be integrated, that can be used by the function \code{Integral}.}

\item{xmin}{\code{numeric} vector lower limits of integration.}

\item{xmax}{\code{numeric} vector upper limits of integration.}

\item{...}{arguments to pass to the function \code{Integral}.}
}
\value{
Return a vector, of the same length as xmin and upper, giving the associated value of the integral.
}
\description{
Extend the Integral function from \code{pracma} by accepting vector for the limits of integration.
}
\examples{
f <- function(x) (x)
xmin <- seq(1,10,length.out = 10)
xmax <- seq(11,20,length.out = 10)
vectIntegral(f, xmin, xmax)

}
