% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectIntegrate.R
\name{vectIntegrate}
\alias{vectIntegrate}
\title{Integrate (vectorized)}
\usage{
vectIntegrate(f, lower, upper, ...)
}
\arguments{
\item{f}{a function to be integrated, that can be used by the function \code{integrate}.}

\item{lower}{\code{numeric} vector lower limits of integration.}

\item{upper}{\code{numeric} vector upper limits of integration.}

\item{...}{arguments to pass to the function \code{integrate}.}
}
\value{
Return a vector, of the same length as lower and upper, giving the associated \code{value} of the integral, other components returned by \code{integrate} are dismissed.
}
\description{
Extend the integrate function from \code{stats} by accepting vector for the limits of integration.
}
\examples{
f <- function(x) (x)
lower <- seq(1,10,length.out = 10)
upper <- seq(11,20,length.out = 10)
}
